- Update readme and code refactor
