+++ src/main/kotlin/com/piashcse/plugins/ConfigureRouting.kt
@@ -22,21 +22,23 @@ fun Application.configureRoute() {
     val orderController: OrderController by inject()
     val paymentController: PaymentController by inject()
     val policyController: PolicyController by inject()
+    val consentController: ConsentController by inject()
     routing {
-        authRoute(authController)
-        profileRoute(userProfileController)
-        shopCategoryRoute(shopCategoryController)
-        shopRoute(shopController)
-        brandRoute(brandController)
-        productCategoryRoute(productCategoryController)
-        productSubCategoryRoute(productSubCategoryController)
-        productRoute(productController)
-        reviewRatingRoute(reviewRatingController)
-        cartRoute(cartController)
-        wishListRoute(wishListController)
-        shippingRoute(shippingController)
-        orderRoute(orderController)
-        paymentRoute(paymentController)
-        policyRoute(policyController)
+        authRoutes(authController)
+        profileRoutes(userProfileController)
+        shopCategoryRoutes(shopCategoryController)
+        shopRoutes(shopController)
+        brandRoutes(brandController)
+        productCategoryRoutes(productCategoryController)
+        productSubCategoryRoutes(productSubCategoryController)
+        productRoutes(productController)
+        reviewRatingRoutes(reviewRatingController)
+        cartRoutes(cartController)
+        wishListRoutes(wishListController)
+        shippingRoutes(shippingController)
+        orderRoutes(orderController)
+        paymentRoutes(paymentController)
+        policyRoutes(policyController)
+        consentRoutes(consentController)
     }
 }
